Simplify hosted runtime startup plumbing
