+++ gestaltd/internal/bootstrap/hosted_workflow_provider.go
@@ -96,8 +96,7 @@ func prepareHostedWorkflowProviderLaunch(ctx context.Context, name string, entry
 		}
 		return nil, fmt.Errorf("query %s support: %w", hostedRuntimeLabel(runtimeConfig), err)
 	}
-	requiresHostnameEgress := deps.Egress.ProviderPolicy(entry).RequiresHostnameEnforcement()
-	runtimePlan := buildRuntimePlacementPlan(runtimeSupport, deps, true, requiresHostnameEgress)
+	runtimePlan := buildRuntimePlacementPlan(runtimeSupport, deps, runtimeRequiresHostnameEgress(entry, deps))
 	if err := runtimePlan.Validate(hostedRuntimeLabel(runtimeConfig)); err != nil {
 		if runtimeOwned {
 			_ = runtimeProvider.Close()
@@ -208,22 +207,9 @@ func startHostedWorkflowProviderInstance(ctx context.Context, launch *hostedWork
 		workflowAllowedHosts = append([]string(nil), launch.allowedHosts...)
 	}
 	allowedHosts := hostedWorkflowAllowedHosts(workflowAllowedHosts, launch.runtimePlan)
-	relayHost := ""
-	if launch.runtimePlan.Resolved.HostServiceAccess == RuntimeHostServiceAccessRelay {
-		bindingEnv, resolvedRelayHost, err := mergeHostedRuntimeHostServiceRelayEnv(name, sessionID, hostServiceBindingDescriptorsFromConfigured(hostServices), deps)
-		if err != nil {
-			return nil, err
-		}
-		relayHost = resolvedRelayHost
-		if len(bindingEnv) > 0 {
-			if startEnv == nil {
-				startEnv = make(map[string]string, len(bindingEnv))
-			}
-			maps.Copy(startEnv, bindingEnv)
-		}
-	}
-	if launch.runtimePlan.RequiresHostnameEgress {
-		allowedHosts = appendAllowedHost(allowedHosts, relayHost)
+	startEnv, allowedHosts, err = applyHostedRuntimeHostServiceRelayEnv(name, sessionID, hostServices, launch.runtimePlan, deps, startEnv, allowedHosts)
+	if err != nil {
+		return nil, err
 	}
 	egressPlan, err := buildHostedRuntimeEgressLaunchPlan(name, sessionID, deps.Egress.Policy(workflowAllowedHosts), allowedHosts, launch.runtimePlan, deps)
 	if err != nil {
